chore: fixed imports of moved non-meteor functions
