+++ imports/plugins/core/catalog/server/methods/publishProducts.js
@@ -1,7 +1,7 @@
 import { Meteor } from "meteor/meteor";
 import { check } from "meteor/check";
 import buildMeteorContext from "/imports/plugins/core/graphql/server/buildMeteorContext";
-import publishProductsMutation from "../mutations/publishProducts";
+import publishProductsMutation from "../no-meteor/mutations/publishProducts";
 
 Meteor.methods({
   "catalog/publish/products": (productIds) => {
