+++ imports/plugins/core/graphql/server/mutations.js
@@ -1,4 +1,4 @@
-import publishProducts from "/imports/plugins/core/catalog/server/mutations/publishProducts";
+import publishProducts from "/imports/plugins/core/catalog/server/no-meteor/mutations/publishProducts";
 
 export default {
   publishProducts
