+++ imports/plugins/core/graphql/server/queries.js
@@ -2,10 +2,10 @@ import { shopAdministratorsQuery } from "/imports/plugins/core/accounts/server/m
 import { userAccountQuery } from "/imports/plugins/core/accounts/server/methods/userAccountQuery";
 import { groupQuery, groupsQuery } from "/imports/plugins/core/accounts/server/methods/groupQuery";
 import { rolesQuery } from "/imports/plugins/core/accounts/server/methods/rolesQuery";
-import catalogItems from "/imports/plugins/core/catalog/server/queries/catalogItems";
-import catalogItemProduct from "/imports/plugins/core/catalog/server/queries/catalogItemProduct";
-import tags from "/imports/plugins/core/catalog/server/queries/tags";
-import tagsByIds from "/imports/plugins/core/catalog/server/queries/tagsByIds";
+import catalogItems from "/imports/plugins/core/catalog/server/no-meteor/queries/catalogItems";
+import catalogItemProduct from "/imports/plugins/core/catalog/server/no-meteor/queries/catalogItemProduct";
+import tags from "/imports/plugins/core/catalog/server/no-meteor/queries/tags";
+import tagsByIds from "/imports/plugins/core/catalog/server/no-meteor/queries/tagsByIds";
 import getShopIdByDomain from "/imports/plugins/core/accounts/server/no-meteor/getShopIdByDomain";
 
 export default {
