+++ imports/plugins/included/inventory/server/methods/inventory.js
@@ -2,7 +2,7 @@ import { Meteor } from "meteor/meteor";
 import { Inventory, Products } from "/lib/collections";
 import { Logger, Reaction } from "/server/api";
 import rawCollections from "/imports/collections/rawCollections";
-import getVariants from "/imports/plugins/core/revisions/server/no-meteor/getVariants";
+import getVariants from "/imports/plugins/core/revisions/server/no-meteor/utils/getVariants";
 
 /**
  * @namespace Inventory/Methods
